case switches, null fixes, recoded a few more things.

+++ source/src/main/java/com/athena/world/content/MoneyPouch.java
@@ -39,7 +39,7 @@ public class MoneyPouch {
 			return false;
 		}
 		if (validateAmount(plr, amount)) {
-			long addedMoney = (long) plr.getMoneyInPouch() + (long) amount;
+			long addedMoney = plr.getMoneyInPouch() + (long) amount;
 			if (addedMoney > Long.MAX_VALUE) {
 				long canStore = Long.MAX_VALUE - plr.getMoneyInPouch();
 				plr.getInventory().delete(5023, (int) canStore);
@@ -77,7 +77,7 @@ public class MoneyPouch {
 			plr.getPacketSender().sendMessage("You cannot do this here.");
 			return false;
 		}
-		long addedMoney = (long) plr.getMoneyInPouch() + (long) amount;
+		long addedMoney = plr.getMoneyInPouch() + amount;
 
 		if (addedMoney > Long.MAX_VALUE) {
 			long canStore = Long.MAX_VALUE - plr.getMoneyInPouch();
